+++ source/PartitionPeriodicGaussian.h
@@ -6,7 +6,7 @@ class PartitionPeriodicGaussian : public Partition {
 public:
 
     PartitionPeriodicGaussian(const std::vector<Param> &params,
-        double period = 2*PI, double images = 10);
+        double period = 2*PI, int images = 10);
 
     // Partition the domain into disjoint intervals
     void partition(double threshold) const;
@@ -20,5 +20,5 @@ public:
 private:
 
     double period_;
-    double images_;
+    int images_;
 };
